+++ NEWS.md
@@ -1,5 +1,9 @@
 # aRtsy 0.1.4
 
+**New features**
+
+- Added `canvas_maze()`.
+
 **Minor changes**
 
 - The function `colorPalette()` now randomizes the `hsl` scale in the `random` color palette instead of the `rgb` scale.